Let MonoReachable traversal evaluate BinOps
